refactor: #53 투표 스케줄러 패키지 및 클래스 추가
